moves some functionality from admin to admin-id
